+++ library/cohesity_job.py
@@ -387,8 +387,8 @@ def _get_tag_ids(module, tags, parentSourceId):
     Function to fetch VMware tag ids for list of tag names.
     """ 
     try:
-        client, status = get_cohesity_client(module)
-        if not status:
+        client = get_cohesity_client(module)
+        if not client:
             module.fail_json(
                 msg="Error while creating cohesity client, err msg '%s'" % client,
                 changed=False)
@@ -416,7 +416,7 @@ def get_vmware_ids(module, job_meta_data, job_details, vm_names):
             str(job_meta_data['parentSourceId'])
         headers = {"Accept": "application/json",
                    "Authorization": "Bearer " + token,
-                   "user-agent": "cohesity-ansible/v2.3.3"}
+                   "user-agent": "cohesity-ansible/v2.3.4"}
         response = open_url(
             url=uri,
             method='GET',
@@ -445,7 +445,7 @@ def get_vmware_vm_ids(module, job_meta_data, job_details, vm_names):
               "/irisservices/api/v1/public/protectionSources/virtualMachines?vCenterId=" + str(job_meta_data['parentSourceId'])
         headers = {"Accept": "application/json",
                    "Authorization": "Bearer " + token,
-                   "user-agent": "cohesity-ansible/v2.3.3"}
+                   "user-agent": "cohesity-ansible/v2.3.4"}
         response = open_url(
             url=uri,
             method='GET',
@@ -485,7 +485,7 @@ def get_view_storage_domain_id(module, self):
         uri = "https://" + server + "/irisservices/api/v1/public/views/" + view_name
         headers = {"Accept": "application/json",
                    "Authorization": "Bearer " + token,
-                   "user-agent": "cohesity-ansible/v2.3.3"}
+                   "user-agent": "cohesity-ansible/v2.3.4"}
         response = open_url(url=uri, method="GET", headers=headers,
                             validate_certs=validate_certs, timeout=REQUEST_TIMEOUT)
         response = json.loads(response.read())
@@ -538,7 +538,7 @@ def register_job(module, self):
         uri = "https://" + server + "/irisservices/api/v1/public/protectionJobs"
         headers = {"Accept": "application/json",
                    "Authorization": "Bearer " + token,
-                   "user-agent": "cohesity-ansible/v2.3.3"}
+                   "user-agent": "cohesity-ansible/v2.3.4"}
         payload = self.copy()
 
         # => Remove the Authorization Token from the Payload
@@ -609,7 +609,7 @@ def start_job(module, self):
             "/irisservices/api/v1/public/protectionJobs/run/" + str(self['id'])
         headers = {"Accept": "application/json",
                    "Authorization": "Bearer " + token,
-                   "user-agent": "cohesity-ansible/v2.3.3"}
+                   "user-agent": "cohesity-ansible/v2.3.4"}
         source_ids = payload.get('sourceIds', [])
         payload = dict()
         payload['runNowParameters'] = [{'sourceId':source_id} for source_id in source_ids]
@@ -657,7 +657,7 @@ def update_job(module, job_details, update_source_ids=None):
             "/irisservices/api/v1/public/protectionJobs/" + str(job_details['id'])
         headers = {"Accept": "application/json",
                    "Authorization": "Bearer " + token,
-                   "user-agent": "cohesity-ansible/v2.3.3"}
+                   "user-agent": "cohesity-ansible/v2.3.4"}
         payload = job_details.copy()
         del payload['token']
         if module.params.get('environment') == 'PhysicalFiles' and module.params.get('delete_sources') == False:
@@ -705,7 +705,7 @@ def get_prot_job_details(self, module):
 
         headers = {"Accept": "application/json",
                    "Authorization": "Bearer " + token,
-                   "user-agent": "cohesity-ansible/v2.3.3"}
+                   "user-agent": "cohesity-ansible/v2.3.4"}
         response = open_url(url=uri, headers=headers,
                             validate_certs=validate_certs, timeout=REQUEST_TIMEOUT)
         if not response.getcode() == 200:
@@ -748,7 +748,7 @@ def stop_job(module, self):
             str(self['id'])
         headers = {"Accept": "application/json",
                    "Authorization": "Bearer " + token,
-                   "user-agent": "cohesity-ansible/v2.3.3"}
+                   "user-agent": "cohesity-ansible/v2.3.4"}
         payload = self.copy()
 
         # => Remove the Authorization Token from the Payload
@@ -800,7 +800,7 @@ def unregister_job(module, self):
             "/irisservices/api/v1/public/protectionJobs/" + str(self['id'])
         headers = {"Accept": "application/json",
                    "Authorization": "Bearer " + token,
-                   "user-agent": "cohesity-ansible/v2.3.3"}
+                   "user-agent": "cohesity-ansible/v2.3.4"}
 
         payload = dict(
             deleteSnapshots=self['deleteSnapshots']