+++ module_utils/storage/cohesity/cohesity_auth.py
@@ -148,7 +148,7 @@ def get__cohesity_auth__token(self):
         auth.username = user_domain[1]
         auth.domain = user_domain[0]
 
-    if "@" in auth.username:
+    elif "@" in auth.username:
         user_domain = auth.username.split("@")
         auth.username = user_domain[0]
         auth.domain = user_domain[1]